Update finalgame.c3p
Adding particles which appear at tail end of rocket when it moves. Making lasers appear at front of rocket using image points. Reworking level 1 to introduce components more slowly. Making dock send player to level 2. However, there is nothing on the level 2 layout so this takes you to an empty screen. Fixing issue where rocket would not appear on screen (this was due to the rocket's position not being reset when the fuel runs out)